Añadidos test unitarios (ok y ko) para la clase PricesService

+++ src/main/java/com/java/testinditex/dto/PriceResponse.java
@@ -3,7 +3,7 @@ package com.java.testinditex.dto;
 import java.math.BigDecimal;
 
 public record PriceResponse(
-        Long productId,
+        Integer productId,
         Integer brandId,
         Long priceList,
         String startDate,
